filesystem storage: don't ensure base path when initialization
The path is only required when storing a package and is guaranteed
before that. Thus it is not needed to ensure that path upon
FileSystemStorage initialization.

This commit also removes the usage of global CONF variable in the
qinling.storage.file_system module, as according the recommendation of
oslo.config it is not recommended.

Change-Id: Id3ccbb172ecc68c135624996b781448e2a69b4a9