kubeadm: fix issue about obtaining k8s config
We use ansible's become to do privilege escalation. However in this way
the $HOME path would be set to /root instead of the user running
devstack. Thus the k8s config would be kept under /root instead of the
$HOME directory of which runs devstack.

This commit fixes the aforementioned issue by directly copying the
config from /etc to the user's directory in setup_gate.sh after running
ansible.

Co-Authored-By: Hunt Xu <mhuntxu@gmail.com>

Change-Id: I8bcffb0f626d91072063e1ce6faced9f7a737a6b