kubernetes: add V1DeleteOptions when deleting services
delete_namespaced_service() now takes an required body (delete option)
parameter in the newly released Kubernetes client 6.0.0 [1]. This is not
backward compatible so this commit fixes that by adding the new required
parameter.

NOTE: the requirement of the kubernetes client will be updated by a
following change to avoid requirements check with this one.

[1] https://github.com/kubernetes-client/python/blob/master/CHANGELOG.md#v600b1

Change-Id: I751691b356bf1ca2e43838bb341e0455286e6a1f