Increase etcd3gw dependency version
There is an upstream issue[1] in etcd3gw, when running on Python3, a
lock is never treated as acquired (although in fact it is acquired).
This makes that qinling cannot work properly on Python3.

The fix[2] is included in etcd3gw 0.2.2. So the requirements must be
updated for qinling to run normally on Python3.

[1]. https://github.com/dims/etcd3-gateway/issues/24
[2]. https://github.com/dims/etcd3-gateway/pull/25

Change-Id: I3868864b957c8a5cd8c26acadc7e17f1c427699c