Retire the Qinling project
As announced in openstack-discuss ML[1], Qinling project
is retiring in Wallaby cycle.

This commit retires this repository as per process deinfed in
project-guide[2]. Anyone would like to maintain it again, please
revert back this commit and propose the re-adding of Qinling
to governance.

The community wishes to express our thanks and appreciation to all of
those who have contributed to the Qinling project over the years.

Depends-On: https://review.opendev.org/c/openstack/project-config/+/764520
Needed-By: https://review.opendev.org/c/openstack/governance/+/764523
[1] http://lists.openstack.org/pipermail/openstack-discuss/2020-November/018638.html
[2] https://docs.openstack.org/project-team-guide/repository.html#retiring-a-repository

Change-Id: Ic9e422dfbba6d23cf83576ece34bb5071cc89377